floating point struggle continues

+++ libft/ft_pow.c
@@ -6,13 +6,13 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/10 10:53:47 by mviinika          #+#    #+#             */
-/*   Updated: 2022/03/01 09:42:50 by mviinika         ###   ########.fr       */
+/*   Updated: 2022/03/24 13:53:12 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_pow(double num, double pow)
+long double	ft_pow(long double num, long double pow)
 {
 	double	res;
 	int		counter;
